Add commit banner
