+++ src/pages/sighting/SightingCore.jsx
@@ -1,7 +1,7 @@
 import React, { useMemo, useState } from 'react';
 import { useHistory } from 'react-router-dom';
 import { useIntl, FormattedMessage } from 'react-intl';
-import { get, flatten, every } from 'lodash-es';
+import { get } from 'lodash-es';
 import { useQueryClient } from 'react-query';
 
 import Tabs from '@material-ui/core/Tabs';
@@ -23,13 +23,11 @@ import Text from '../../components/Text';
 import LoadingScreen from '../../components/LoadingScreen';
 import SadScreen from '../../components/SadScreen';
 import Button from '../../components/Button';
-import Alert from '../../components/Alert';
 import MoreMenu from '../../components/MoreMenu';
 import ConfirmDelete from '../../components/ConfirmDelete';
 import EntityHeaderNew from '../../components/EntityHeaderNew';
 import useDocumentTitle from '../../hooks/useDocumentTitle';
 import useDeleteSighting from '../../models/sighting/useDeleteSighting';
-import useCommitAssetGroupSighting from '../../models/assetGroupSighting/useCommitAssetGroupSighting';
 import useSightingFieldSchemas from '../../models/sighting/useSightingFieldSchemas';
 import { formatDate } from '../../utils/formatters';
 // import AnnotationsGallery from './AnnotationsGallery';
@@ -38,6 +36,7 @@ import Annotations from './Annotations';
 import Photographs from './Photographs';
 import OverviewContent from './OverviewContent';
 import SightingHistoryDialog from './SightingHistoryDialog';
+import CommitBanner from './CommitBanner';
 import FeaturedPhoto from './featuredPhoto/FeaturedPhoto';
 import Encounters from './encounters/Encounters';
 
@@ -68,11 +67,6 @@ export default function SightingCore({
     error: deleteSightingError,
     setError: setDeleteSightingError,
   } = useDeleteSighting();
-
-  const {
-    commitAgs,
-    isLoading: commitAgsLoading,
-  } = useCommitAssetGroupSighting();
 
   /*
     known issue: if data or fieldschemas change values
@@ -116,13 +110,6 @@ export default function SightingCore({
   if (error) return <SadScreen variant="genericError" />;
 
   const assets = get(data, 'assets', []);
-  const annotations = flatten(
-    assets.map(asset => get(asset, 'annotations')),
-  );
-  const allAnnotationsAssigned = every(annotations, 'encounter_guid');
-  const detectionComplete = get(data, 'stage') !== 'detection';
-  const showCommitAlert =
-    pending && detectionComplete && allAnnotationsAssigned;
 
   const sightingDisplayDate = get(data, 'time');
 
@@ -222,22 +209,11 @@ export default function SightingCore({
           </Text>
         )}
       </EntityHeaderNew>
-      {showCommitAlert && (
-        <Alert
-          severity="info"
-          titleId="SIGHTING_COMMIT_TITLE"
-          descriptionId="SIGHTING_COMMIT_DESCRIPTION"
-          action={
-            <Button
-              id="COMMIT"
-              size="small"
-              loading={commitAgsLoading}
-              onClick={() => commitAgs(id)}
-              style={{ marginTop: 8, marginRight: 4 }}
-            />
-          }
-        />
-      )}
+      <CommitBanner
+        sightingId={id}
+        pending={pending}
+        sightingData={data}
+      />
       {activeTab === '#overview' && (
         <OverviewContent
           metadata={metadata}
